+++ app/helpers/camaleon_cms/theme_helper.rb
@@ -3,6 +3,8 @@ module CamaleonCms
     def theme_init
       breadcrumb_items = []
       camaleon_theme_state[:front_breadcrumb] = breadcrumb_items
+      # back-compat ivar: themes/plugins may read @_front_breadcrumb directly
+      instance_variable_set(:@_front_breadcrumb, breadcrumb_items)
     end
 
     # return theme full asset path